Generate 50 euromillions keys
Create a Python program that generates 50 keys for the euro-millions draw. This is made up of a sequence of 5 numbers from 1 to 50 and another sequence of 2 numbers from 1 to 12. There cannot be repeated numbers in each sequence.
Then the program must save the keys in a csv file. The csv file header should have the format:
number1, number2, number3, number4, number5, star1, star2
Of these 50 keys, make a graph with the number of times the numbers and stars were generated.