Python program that creates a class with attributes
Write a Python program in which you create a class with attributes and methods so that:

Create name, number, age, class as attributes
Create methods to obtain attributes
Allow the user to consult existing students and their data.
Create an interface that allows the user to select the option they want, always running. Only with a specific option ends the program
Students are created by the user in the program.